Add generic f64 support
By using the num-traits crate the easing functions
can rely on num_traits::Float rather than f32, as
such all functions will support f32 & f64 precision.
Usage of literals in the equations had to be
converted to cast to the generic type.
Also added type docs to the Easing trait, to be
helpful when looking at the docs.
Reworked example graph to be a little simpler.